# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/travel-plans/{travelPlanId}/participant/{participantId}/expenses/{expenseId}": {
      "put": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "updateExpense",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "deleteExpense",
        "parameters": [
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/travel-plans/{travelPlanId}/participant/{participantId}/documents/{documentId}": {
      "put": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "updateDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "deleteDocument",
        "parameters": [
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/document-types/{id}": {
      "put": {
        "tags": [
          "document-type-controller"
        ],
        "operationId": "update_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "document-type-controller"
        ],
        "operationId": "delete_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/travel-plans": {
      "get": {
        "tags": [
          "travel-plan-controller"
        ],
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListTravelPlanSummaryResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "travel-plan-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTravelPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/travel-plans/{travelPlanId}/participant/{participantId}/expenses": {
      "post": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "createExpense",
        "parameters": [
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/travel-plans/{travelPlanId}/participant/{participantId}/documents": {
      "post": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "createDocument",
        "parameters": [
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/document-types": {
      "get": {
        "tags": [
          "document-type-controller"
        ],
        "operationId": "getAll_6",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListDocumentTypeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "document-type-controller"
        ],
        "operationId": "create_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/travel-plans/{travelPlanId}/participant/{participantId}/expenses/{expenseId}/submit": {
      "patch": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "submitExpense",
        "parameters": [
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/travel-plans/{travelPlanId}/participant/{participantId}": {
      "get": {
        "tags": [
          "travel-plan-participant-controller"
        ],
        "operationId": "getParticipant",
        "parameters": [
          {
            "name": "travelPlanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultParticipantResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {

          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateExpenseRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "travelPlanId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "float",
            "minimum": 1
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "expenseCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "proofs": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        },
        "required": [
          "amount",
          "expenseCategoryId",
          "id",
          "travelPlanId"
        ]
      },
      "UpdateDocumentRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "doc": {
            "type": "string",
            "format": "binary"
          },
          "travelPlanId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "documentTypeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "doc",
          "documentTypeId",
          "id",
          "ownerId",
          "travelPlanId"
        ]
      },
...
      "UpdateDocumentTypeRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "CreateTravelPlanRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "destination": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "destination",
          "title"
        ]
      },
      "CreateExpenseRequest": {
        "type": "object",
        "properties": {
          "participantId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "float",
            "minimum": 1
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFTING",
              "SUBMITTED",
              "APPROVED",
              "REJECTED"
            ]
          },
          "travelPlanId": {
            "type": "string",
            "format": "uuid"
          },
          "expenseCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "proofs": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        },
        "required": [
          "amount",
          "expenseCategoryId",
          "travelPlanId"
        ]
      },
      "CreateDocumentRequest": {
        "type": "object",
        "properties": {
          "doc": {
            "type": "string",
            "format": "binary"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "travelPlanId": {
            "type": "string",
            "format": "uuid"
          },
          "documentTypeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "doc",
          "documentTypeId",
          "ownerId",
          "travelPlanId"
        ]
      },
...
      "CreateDocumentTypeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
...
      "GlobalUserResponseSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfileResponseSummary"
          }
        }
      },
      "ResultListTravelPlanSummaryResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TravelPlanSummaryResponse"
            }
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "TravelPlanSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "updatedBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          }
        }
      },
      "UserProfileResponseSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        }
      },
      "ExpenseProofResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "docUrl": {
            "type": "string"
          }
        }
      },
      "ParticipantDocumentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "docUrl": {
            "type": "string"
          },
          "documentType": {
            "type": "string"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "uploadedBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          }
        }
      },
      "ParticipantExpenseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFTING",
              "SUBMITTED",
              "APPROVED",
              "REJECTED"
            ]
          },
          "remarks": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expenseCategory": {
            "type": "string"
          },
          "approvedBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "proofs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseProofResponse"
            }
          }
        }
      },
      "ParticipantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantDocumentResponse"
            }
          },
          "expenses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantExpenseResponse"
            }
          }
        }
      },
      "ResultParticipantResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/ParticipantResponse"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "DocumentTypeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "updatedBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultListDocumentTypeResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentTypeResponse"
            }
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}